backend-engineer/microservices/product_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
redis
python-dotenv
